requirements.txt
fastapi==0.103.0
passlib==1.7.4
psycopg2==2.9.5
PyJWT==2.8.0
bcrypt==4.0.1